Problem 2872: Improved solution. Temp. switch to java to focus on the algorithm, instead of learning Rust. Took me a long time, before it clicked due to glazing over the fact it is a undirected tree !
Solution is now changed to a "merge or split" algorithm.

The idea is that no leaf (edges = 1) can exist if "value % k != 0".
If the value is divisible by 0, just split! (There is no better solution than eagerly splitting).
Else merge the leaf with the "parent" (combining the value).

The solution does this until there are no leafs anymore (except for the root).

O(n^2) due to how it is implemented. Can easily adjust it to O(n).